+++ AssetsFlow.Client.React/CHANGELOG.md
@@ -3,6 +3,8 @@
 
 ## [0.9.017-c] - 2025-06-22
 ### Changed
+- Journal loading states
+- Dev indicators (browser and page headers)
 
 ## [0.9.016-c] - 2025-06-22
 ### Changed